+++ db/migrate/20230626084707_add_conversion_risk_protection_arrangement_task_attributes.rb
@@ -1,0 +1,5 @@
+class AddConversionRiskProtectionArrangementTaskAttributes < ActiveRecord::Migration[7.0]
+  def change
+    add_column :conversion_tasks_data, :risk_protection_arrangement_option, :string
+  end
+end